requirements.txt
streamlit>=1.35.0
openai>=1.35.0
python-dotenv>=1.0.1
PyPDF2>=3.0.1
python-docx>=1.1.0
docxtpl>=0.16.7